interfaces/api/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
httpx==0.28.1
pyyaml==6.0.2
